Добавил кэш запросов

+++ app/src/main/java/com/example/e_vasiliev/moneyconverter/utils/Utils.java
@@ -12,6 +12,7 @@ import android.view.View;
 import android.view.inputmethod.InputMethodManager;
 import android.widget.Toast;
 
+import com.example.e_vasiliev.moneyconverter.network.retrofit.models.ConverterOutput;
 import com.example.e_vasiliev.moneyconverter.network.retrofit.models.CurrencyModel;
 import com.google.gson.Gson;
 
@@ -21,6 +22,7 @@ import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.io.OutputStreamWriter;
+import java.util.Calendar;
 import java.util.regex.Pattern;
 
 public final class Utils {
@@ -42,22 +44,24 @@ public final class Utils {
         return netInfo != null && netInfo.isConnectedOrConnecting();
     }
 
+
     /**
      * @param layout  - view, в которой будет расположен snackbar. Предпочтительнее {@link android.support.design.widget.CoordinatorLayout}
      * @param message - строковая контстанта, которая будет отображена в сообщении
      */
     public static void showMessage(View layout, @StringRes int message) {
-        Snackbar.make(layout, message, Snackbar.LENGTH_SHORT).show();
+        Snackbar.make(layout, message, Snackbar.LENGTH_LONG).show();
     }
 
 
-    public static void hideKeyboard(Activity activity){
+    public static void hideKeyboard(Activity activity) {
         InputMethodManager inputManager = (InputMethodManager)
                 activity.getSystemService(Context.INPUT_METHOD_SERVICE);
 
         inputManager.hideSoftInputFromWindow(activity.getCurrentFocus().getWindowToken(),
                 InputMethodManager.HIDE_NOT_ALWAYS);
     }
+
 
     public static boolean checkText(CharSequence text) {
         if (pattern.matcher(text).find()) {
@@ -67,7 +71,44 @@ public final class Utils {
     }
 
 
-    public static CurrencyModel getFromCache(Context context) {
+    public static void saveConverterOutput(Context context, ConverterOutput converterOutput, String id) {
+        Log.d("converter", "пытаюсь сохранить ConverterOutput в кэш");
+        if (converterOutput.getDatetime() == null) {
+            converterOutput.setDatetime(Calendar.getInstance().getTime());
+        }
+
+        Gson gson = new Gson();
+        String json = gson.toJson(converterOutput);
+
+        deleteFileIfExists(context, id);
+        createFile(context, id);
+        writeInFile(context, id, json);
+    }
+
+
+    public static ConverterOutput getConverterOutputFromCache(Context context, String id) {
+        Log.d("converter", "пытаюсь получить ConverterOutput из кэша");
+        byte[] buffer = null;
+        try {
+            FileInputStream inputStream = context.openFileInput(id);
+            buffer = new byte[inputStream.available()];
+            inputStream.read(buffer, 0, buffer.length);
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+
+        if (buffer != null) {
+            String json = new String(buffer);
+            Gson gson = new Gson();
+            return gson.fromJson(json, ConverterOutput.class);
+        }
+
+
+        return null;
+    }
+
+
+    public static CurrencyModel getCurrencyFromCache(Context context) {
         byte[] buffer = null;
         try {
             FileInputStream inputStream = context.openFileInput(filename);
@@ -87,21 +128,16 @@ public final class Utils {
     }
 
 
-    public static void saveInCache(Context context, CurrencyModel currencyModels) {
-
+    public static void saveCurrencyInCache(Context context, CurrencyModel currencyModels) {
         Gson gson = new Gson();
         String json = gson.toJson(currencyModels);
-        File file = new File(context.getCacheDir(), filename);
-        if (file.exists()) {
-            deleteFile(file);
-        } else {
-        }
-
+        deleteFileIfExists(context, filename);
         createFile(context, filename);
         writeInFile(context, filename, json);
     }
 
 
+    //region манипуляции с файлом
     private static File createFile(Context context, String filename) {
         File file = null;
         try {
@@ -118,14 +154,17 @@ public final class Utils {
             OutputStream outputStream = context.openFileOutput(filename, Context.MODE_PRIVATE);
             outputStream.write(data.getBytes());
             outputStream.close();
-
         } catch (IOException e) {
             e.printStackTrace();
         }
     }
 
 
-    private static void deleteFile(File file) {
-        file.delete();
+    private static void deleteFileIfExists(Context context, String filename) {
+        File file = new File(context.getCacheDir(), filename);
+        if (file.exists()) {
+            file.delete();
+        }
     }
+    //endregion
 }
